requirements.txt
openai>=1.0.0
tiktoken>=0.5.0
chromadb>=0.4.0
Pillow>=9.0.0
flask>=2.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
gunicorn>=21.0.0